+++ udata_front/views/schema.py
@@ -2,9 +2,9 @@ from urllib.parse import urlencode
 
 from flask import Blueprint, current_app
 from udata.app import cache
+from udata.frontend import template_hook
 
 from udata_front import theme
-from udata_front.frontend import template_hook
 
 import requests