+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Move template hook logic to udata and add oauth hooks [#29](https://github.com/etalab/udata-front/pull/29)
 
 ## 1.1.1 (2021-10-22)